Renaming Package names to io.opentelemetry.sqlcommenter
