+++ java/sqlcommenter-java/build.gradle
@@ -9,7 +9,7 @@ apply plugin: "net.ltgt.errorprone"
 apply plugin: "jacoco"
 apply plugin: "signing"
 
-group = "com.google.cloud"
+group = "io.opentelemetry"
 version = "2.0.0" // CURRENT_VERSION
 
 sourceCompatibility = 1.8
@@ -140,53 +140,3 @@ task sourcesJar(type: Jar) {
 artifacts {
     archives javadocJar, sourcesJar
 }
-
-uploadArchives {
-        repositories {
-            mavenDeployer {
-                beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
-
-                def configureAuth = {
-                    if (rootProject.hasProperty('ossrhUsername') && rootProject.hasProperty('ossrhPassword')) {
-                        authentication(userName:rootProject.ossrhUsername, password: rootProject.ossrhPassword)
-                    }
-                }
-
-                repository(url: "https://oss.sonatype.org/service/local/staging/deploy/maven2/", configureAuth)
-
-                snapshotRepository(url: "https://oss.sonatype.org/content/repositories/snapshots/", configureAuth)
-
-                pom.project {
-                    name "sqlcommenter-hibernate"
-                    packaging 'jar'
-                    description project.description
-                    url 'https://github.com/open-telemetry/opentelemetry-sqlcommenter'
-
-                    scm {
-                        connection 'scm:svn:https://github.com/open-telemetry/opentelemetry-sqlcommenter'
-                        developerConnection 'scm:git:git@github.com/google/sqlcommenter'
-                        url 'https://github.com/open-telemetry/opentelemetry-sqlcommenter'
-                    }
-
-                    licenses {
-                        license {
-                            name 'The Apache License, Version 2.0'
-                            url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-                        }
-                    }
-
-                    developers {
-                        developer {
-                            id 'com.google.cloud'
-                            name 'Google Developers'
-                            email 'sqlcommenter@googlegroups.com'
-                            url 'cloud.google.com'
-                            // https://issues.gradle.org/browse/GRADLE-2719
-                            organization = 'SQLCommenter Authors'
-                            organizationUrl 'https://google.com'
-                        }
-                    }
-                }
-            }
-        }
-    }
