+++ java/sqlcommenter-java/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = "sqlcommenter"
+rootProject.name = "opentelemetry-sqlcommenter"
