--- java/sqlcommenter-java/src/test/java/com/google/cloud/sqlcommenter/spring/backend/dao/PostRepository.java
@@ -1,26 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.cloud.sqlcommenter.spring.backend.dao;
-
-import com.google.cloud.sqlcommenter.spring.backend.domain.Post;
-import java.util.List;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends CrudRepository<Post, Long> {
-
-  List<Post> findByTitle(String title);
-}
