--- java/sqlcommenter-java/src/test/java/com/google/cloud/sqlcommenter/spring/backend/domain/Post.java
@@ -1,55 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.cloud.sqlcommenter.spring.backend.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.persistence.*;
-
-@Entity
-@Table(name = "post")
-public class Post {
-
-  @Id @GeneratedValue private Long id;
-
-  private String title;
-
-  @ManyToMany
-  @JoinTable(
-      name = "post_tag",
-      joinColumns = @JoinColumn(name = "post_id"),
-      inverseJoinColumns = @JoinColumn(name = "tag_id"))
-  private List<Tag> tags = new ArrayList<>();
-
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getTitle() {
-    return title;
-  }
-
-  public void setTitle(String title) {
-    this.title = title;
-  }
-
-  public List<Tag> getTags() {
-    return tags;
-  }
-}
